requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
nbformat
nltk
openai
fonttools==4.60.1
frozendict==2.4.6
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
multitasking==0.0.12
numpy==2.3.3
python-dotenv
packaging==25.0
pandas==2.3.3
pdfminer.six==20240706
peewee==3.18.2
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.32.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn
seaborn
six==1.17.0
soupsieve==2.8
textblob
torch
transformers
tqdm
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xgboost
yfinance==0.2.66
